Refactor LeasesCard
