--- src/features/customerView/leasesCard/LeasesCard.test.tsx
@@ -1,51 +1,0 @@
-import React from 'react';
-import { mount } from 'enzyme';
-import { HashRouter } from 'react-router-dom';
-
-import LeasesCard, { LeasesCardProps } from './LeasesCard';
-import { LeaseStatus } from '../../../@types/__generated__/globalTypes';
-
-describe('LeasesCard', () => {
-  const initialProps: LeasesCardProps = {
-    title: 'title',
-    infoSectionTitle: 'infoSectionTitle',
-    addressLabel: 'addressLabel',
-    leaseDetails: [
-      {
-        id: '1234',
-        address: 'Pursilahdenranta',
-        startDate: '2019-06-14',
-        endDate: '2019-09-10',
-        link: 'link',
-        status: LeaseStatus.DRAFTED,
-      },
-      {
-        id: '4321',
-        address: 'Telakkakatu 1',
-        startDate: '2019-05-20',
-        endDate: '2020-06-15',
-        link: 'link',
-        status: LeaseStatus.TERMINATED,
-      },
-    ],
-    cancelLease: jest.fn(),
-    customerName: 'Tester McTestington',
-  };
-
-  const getWrapper = (props?: Partial<LeasesCardProps>) =>
-    mount(
-      <HashRouter>
-        <LeasesCard {...initialProps} {...props} />
-      </HashRouter>
-    );
-
-  beforeEach(() => {
-    jest.restoreAllMocks();
-  });
-
-  it('renders normally', () => {
-    const wrapper = getWrapper();
-
-    expect(wrapper.render()).toMatchSnapshot();
-  });
-});
