+++ src/features/customerView/leasesCard/WinterStorageLeasesCard.tsx
@@ -1,9 +1,12 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 
-import { WinterStorageLease } from '../types';
-import LeasesCard from './LeasesCard';
 import WinterStorageContractDetailsContainer from '../../contractDetails/WinterStorageContractDetailsContainer';
+import { Lease } from './types';
+import { WinterStorageLease } from '../types';
+import Card from '../../../common/card/Card';
+import CardHeader from '../../../common/cardHeader/CardHeader';
+import LeaseDetails from './leaseDetails/LeaseDetails';
 
 export interface WinterStorageLeasesCardProps {
   cancelLease: (id: string, type: 'berth' | 'winterStorage') => void;
@@ -13,27 +16,33 @@ export interface WinterStorageLeasesCardProps {
 
 const WinterStorageLeasesCard = ({ customerName, cancelLease, leases }: WinterStorageLeasesCardProps) => {
   const { t } = useTranslation();
-  const leaseDetails = leases.map((lease) => {
+  const mapLeaseDetails = (lease: WinterStorageLease): Lease => {
     return {
       address: lease.winterStorageArea?.name || '',
       endDate: lease.endDate,
       id: lease.id,
       link: lease?.winterStorageArea?.id ? `/winter-storage-areas/${lease?.winterStorageArea?.id}` : undefined,
-      renderContractDetails: (leaseId: string) => <WinterStorageContractDetailsContainer leaseId={leaseId} />,
       startDate: lease.startDate,
       status: lease.status,
+      renderContractDetails: () => <WinterStorageContractDetailsContainer leaseId={lease.id} />,
     };
-  });
+  };
 
   return (
-    <LeasesCard
-      addressLabel={t('customerView.leases.winterStorage.addressLabel')}
-      customerName={customerName}
-      cancelLease={(id) => cancelLease(id, 'winterStorage')}
-      infoSectionTitle={t('customerView.leases.winterStorage.infoSectionTitle')}
-      leaseDetails={leaseDetails}
-      title={t('customerView.leases.winterStorage.title')}
-    />
+    <Card>
+      <CardHeader title={t('customerView.leases.winterStorage.title')} />
+      {leases.map((lease) => (
+        <LeaseDetails
+          key={lease.id}
+          addressLabel={t('customerView.leases.winterStorage.addressLabel')}
+          cancelLease={(id) => cancelLease(id, 'winterStorage')}
+          createLease={() => undefined}
+          customerName={customerName}
+          infoSectionTitle={t('customerView.leases.winterStorage.infoSectionTitle')}
+          {...mapLeaseDetails(lease)}
+        />
+      ))}
+    </Card>
   );
 };
