+++ src/features/customerView/leasesCard/leaseDetails/leaseDetails.module.scss
