+++ src/features/customerView/leasesCard/types.ts
@@ -1,0 +1,17 @@
+import React from 'react';
+
+import { BerthMooringType, LeaseStatus } from '../../../@types/__generated__/globalTypes';
+
+export type Lease = {
+  address: string;
+  depth?: number | null;
+  endDate: string;
+  id: string;
+  length?: number;
+  link?: string;
+  mooringType?: BerthMooringType;
+  startDate: string;
+  status: LeaseStatus;
+  width?: number;
+  renderContractDetails(): React.ReactNode;
+};
